+++ src/main/java/org/folio/rs/domain/entity/ProviderRecord.java
@@ -1,8 +1,12 @@
 package org.folio.rs.domain.entity;
 
+import lombok.Getter;
+
+@Getter
 public enum ProviderRecord {
   DEMATIC_EMS("Dematic EMS", "DEMATIC_EMS"),
-  DEMATIC_SD("Dematic StagingDirector", "DEMATIC_SD");
+  DEMATIC_SD("Dematic StagingDirector", "DEMATIC_SD"),
+  CAIA_SOFT("CaiaSoft", "CAIA_SOFT");
 
   private final String name;
   private final String id;
@@ -10,13 +14,5 @@ public enum ProviderRecord {
   ProviderRecord(String name, String id) {
     this.name = name;
     this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public String getId() {
-    return id;
   }
 }
